requirements.txt
# Core ML libraries
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
numpy==1.26.2
pandas==2.2.0
scikit-learn==1.4.0
tqdm==4.66.1

# HuggingFace for BLIP & Wav2Vec2
transformers==4.36.2
datasets==2.15.0

# Torch-Geometric (ST-GNN)
torch-geometric==2.4.0
torch-scatter==2.1.2
torch-sparse==0.6.18
torch-cluster==1.6.3
torch-spline-conv==1.2.2

# FastAPI server
fastapi==0.109.0
uvicorn==0.27.0

# Monitoring
prometheus-client==0.19.0

# Visualization
matplotlib==3.8.2

# Kubernetes interaction
kubernetes==29.0.0

# Others
pydantic==2.5.0
requests==2.31.0
Pillow==10.1.0
